requirements.txt
requests~=2.31.0
python-dotenv~=1.0.1
loguru~=0.7.2
watchdog~=4.0.0
black~=24.3.0
flake8~=7.0.0